refactor(html): move metadata?
