+++ void-html/src/main/kotlin/io/void/html/util/HtmlResponse.kt
@@ -1,11 +1,9 @@
 package io.void.html.util
 
 import io.void.dto.http.ResponseDTO
-import io.void.dto.http.buildResponse
-import io.void.dto.http.headers
 import io.void.dto.http.ok
 import io.void.html.Element
-import io.void.html.page.metadata.Metadata
+import io.void.html.metadata.Metadata
 
 fun createResponse(element: Element, metadata: Metadata): ResponseDTO = ok("""
                         <!doctype html><html>
